Remove useless method from Writer class
